Renamed `parsl.jobs.simple_error_handler` to `parsl.jobs.error_handlers`
