+++ parsl/executors/status_handling.py
@@ -10,7 +10,7 @@ import parsl  # noqa F401
 from parsl.executors.base import ParslExecutor
 from parsl.executors.errors import BadStateException, ScalingFailed
 from parsl.jobs.states import JobStatus, JobState
-from parsl.jobs.simple_error_handler import simple_error_handler, noop_error_handler
+from parsl.jobs.error_handlers import simple_error_handler, noop_error_handler
 from parsl.providers.base import ExecutionProvider
 from parsl.utils import AtomicIDCounter
